python_modules/requirements.txt
# Python dependencies for OSINT modules
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyGithub>=1.59.0
shodan>=1.31.0
censys>=2.2.0
